Let slaves reconnect to master on errors
Before whenever a slave got an errors returned from the master it was
only logged. With this same it assumes that on error the slave becomes
useless and reconnects to the master.

Todo: better error handling on the master side, better handling of the
remaining data on the slaves side.