go-fuzz: switch to a PCG PRNG
go-fuzz generates a lot of random numbers.
math/rand is a mediocre PRNG, is large, and has contention issues.
This is mostly the fault of the Go 1 compatibility promise.

Add a dedicated PRNG just for go-fuzz; it's just 100 lines of code.
Because a 64 bit period and 32 bit output is ample for go-fuzz,
we can use a small PCG generator.
This is higher quality and much faster:
This PCG's Uint32 method is 2x faster than math/rand,
and the Intn method is 3x faster.

It is also easier to use.