go-fuzz: use a re-usable buffer for testee communication
For a fast-running Fuzz function, we might restart a testee
multiple times a second. This giant allocation gets expensive
in that scenario.

It is also a significant slowdown once go-fuzz starts finding crashers.

Makes testee restarts cheaper is also an important step
towards supporting Fuzz functions that are not idempotent,
e.g. because they are hacked into an existing large code base,
such as cmd/compile.

Instead of allocating 1mb on every restart, allocate a single
re-usable buffer up front. We don't even need to clear it between runs,
which makes this particularly cheap and effective.