go-fuzz-build: improve pkg resolution
Prior to this change, go-fuzz-build assumed that the
first package in the slice returned by packages.Load
corresponded to the package requested by the user.
This assumption was faulty, although it worked accidentally.

This change adds another package.Load call to explicitly
resolve the package to be fuzzed, and does more careful validation of it.

While we're here, do some minor cleanup:

* Request cheaper loading of std. All we need is package paths.
* Simplify processing of the result of loading std.
* Store the fuzz package, rather than its package path, in Context.
  This will be helpful in other planned work (e.g. adding fuzz.F).

Updates golang/go#30826 and golang/go#30828