go-fuzz-dep: set input capacity to len
The input capacity was always set MaxInputSize (1MB).
This means that go-fuzz could miss whole lot of out-of-bounds bugs
related to slicing, since slicing operations look at capacity.
E.g. calculating indices of a field in the input and then
extracting the field with slicing. Somehow we missed this for
all this time.
Found thanks to libfuzzer mode that does this properly and
fuzzit.dev that uses the libfuzzer mode.