Add execution duration flag
Add flag for automatic shutdown after a specified duration.

It seems there can be advantages of restarting fuzzing at regular intervals as seen in #257

Also it can be useful to run fuzz tests for a short period to check for regressions.

To facilitate this, add a `-duration` parameter. After the fuzzer has been running this duration it will do a regular exit.

This will always exit with code 0. It could make sense to exit with non-zero if there it one or more crashers.