go-fuzz-build: do not instrument package reflect
This is unfortunate, but it circumvents a build failure:

failed to execute go build: exit status 1
# reflect
/Users/josh/go/1.21/src/reflect/value.go:3958: misplaced compiler directive

The problem is that in this code:

//go:nosplit
func noescape(p unsafe.Pointer) unsafe.Pointer {

The nosplit gets separated from the func, and placed
in the body of the function above it. Oops.

We could probably be more careful and make this work,
but the fix is non-obvious, and probably not even in this package.

This project is on life support.
Ignoring reflect is the simplest possible fix.